I've added in a SIRE_USE_OPENMM option that will allow users to disable use of OpenMM even if it is detected on their system.
In doing this I have had to create a Move/CMakeOpenMMFile.txt that
separates out all of the OpenMM-dependent wrappers and put them
into a different file. I have edited the CMake files to only
pull in OpenMM files if needed. This now all compiles and links
on my MacBook (Yosemite)


Former-commit-id: bc25ec94870cad90c8ca70c80255f9b52724bfc6