I've added a user-stats uploader that will upload some information to siremol.org about the Sire every time it is run. This can be disabled by setting the environment variable
SIRE_DONT_PHONEHOME=1

This tries to upload CPU, Sire version, executable and
git repository information to siremol.org, but does not
include any user-identifiable information. The upload occurs
in a background thread and will exit without issue if there
are any problems. You can see what information has been uploaded
by inspecting the Sire.sent_usage_data variable after the
upload has completed. An example is;

{'version': '2015.0.0', 'clockspeed': 2294, 'repository_version': 'cf9cfea2507d9f80efdd9240413e1166bf933a45 [formerly f46da231489a2e08d615e70f12a38554ea20ba37]/Christopher Woods/2015-09-10T17:55:27+01:00', 'executable': 'ligandswap', 'processor': 'Intel(R) Core(TM) i5-3427U CPU @ 1.80GHz', 'vendor': 'GenuineIntel', 'repository': 'https://github.com/michellab/Sire.git | merge_2014_4', 'numcores': 4}

This information will be useful for Sire developers to see how often
the code is being used, which processors and version need to be
supported, and which of the Sire apps are being used.


Former-commit-id: b488a122f4d9e4453f3450a90fa5d530978a204a