I've now significantly sped up the adding of coordinates/velocities to molecular systems loaded from a gromacs top file. I did this by adding in a renumber function to molecule, so that we don't have to go to and from a MolStructureEditor to complete the task of simply renumbering atoms or residues. This also fixed a minor bug in which renumbering would affect the molecule, but any info objects in the property were not renumbered.
With this, loading a gromacs top/gro87 file of a fully solvated
protein-ligand system takes about 3 seconds on my macbook, which
is acceptably quick. The tests are all passing, so this
can be merged into devel. The next step is to use this
reader/writer with the gromacs tutorial to see if we can
actually get gromacs to read and understand the files that
Sire is writing...

[ci skip]


Former-commit-id: ee3ab9cce673a58a6f39bdd2d18ca5dd589a95fd