I've fixed the AmberPrm scl14 writing bug, and the SOLTY writing bug.
Also removed the background process printing to screen for the build as
this is no longer needed, and was causing weird error message on
OS X Mojave.

Now, if the 14 scale factor is not set in an Amber input it will
automatically set the scale factor to 0, thereby ignoring the
CLJ interactions completely. Comparison to existing behaviour
and gromacs converted topologies suggests this is the correct
course of action.


Former-commit-id: 70783db295f9fdfd5c7de4f166813ac2284e1858