Working on letting AmberPrm process PARM7 files that include SS bonds, or other bonds between two "separate" molecules. This has involved refactoring the logic as suggested in issue #236 and #217 (thanks Paolo).
Currently this can read the molecules and pass most tests. However
still have failures for an SS bond, as need to fully propogate the
change through the code.

Also have failure for PRM7 files that don't have bonds for water
molecules. This is because we find molecules by scanning
the connectivity, so water molecules without bonds are coming
up as three 1-atom molecules (which is triggering a sanity
assertation comparing what ATOMS_PER_MOL in the file is saying
versus what we are seeing. The solution is to likely use the
starting solvent index in the file to stop the bond checking for
solvent molecules.

While this fails, I will still do a CI/CD to ensure that the
pipeline will notify me of test failures, and will not deploy.


Former-commit-id: 6a88c8988610d04a40eb92eab95f73bd3f57da81