I've merged in everything from devel as described in my message in the pull request.
Everything compiles and runs. The new pytest tests pass, as do all of the sire_test
tests (except for the one expected failure in SireMol due to the change in the way
that selections operate)

Once pushed, I will check the pull request to see whether this has resolved the
conflicts.


Former-commit-id: 8c24f178846b848db8ee530772ecdeee3e62f68a