Fixed the bug where assigning a list or dictionary to an atom property resulted in an AttributeError (no typeName function).
In doing this (and writing the tests) I also needed to add in code that
smooths the auto-conversion / comparison between base type wrapped
properties (list NumberProperty) to the base type (e.g. int or float).

As much as is possible, user's should feel that they are working directly
with native python types now, and code like this works as expected

>>> cursor = mol.cursor()[0]
>>> cursor["atom_test"] = ["dog", 3.1, 5]
>>> mol = cursor.parent().commit()
>>> assert mol[0].property("atom_test")[0] == "dog"
>>> assert mol[0].property("atom_test")[1] == 3.1
>>> assert mol[0].property("atom_test")[2] == 5

(this code is in the new 'test_cursor_assign_from_container' test)


Former-commit-id: 244bca454f8d76ac0f946e8f6df4ee2e67eb2e6f