Added the ability to search on whether or not a property exists.
This will now find views that have the property, if the property
is not equal to zero (false). The new test_property_search.py
tests this, and demonstrates how it can be used. This hasn't
changed the search syntax, so works as would be anticipated,
e.g.

>>> mols["property charge"]

returns all molecules that have a defined charge property.


Former-commit-id: 941603ba2dd5ef8fa39e8859e364c6fc8858bd8f