Added full 'map' support to searches. This should make integration with BioSimSpace easier, as now it can directly use the searches and can pass in the maps too :-)
For example;

>>> mols.atoms("element C", map=map)

>>> mols[("element C", map)]

>>> mols.search("element C", map=map)

are all equivalent.

If you pass in a list of tuple of length 2 to __getitem__ where the
second item can be converted to a propertymap using sr.base.create_map(),
then the search term will be the first item, and the property map
will be the second item.

I've tried this with some searches, e.g.

>>> mols[("atom mass > 5", {"mass": "charge"})]
Error - cannot convert a charge to a mass

>>> mols[("atoms within 15.0 of molidx -2", {"space": sr.vol.Cartesian()})]
864 atoms

>>> mols["atoms within 15.0 of molidx -2"]
1400 atoms (uses periodic box of the molecules)

>>> mols[("atoms within 15.0 of molidx -2", {"space": mols.property("space")})]
1400 atoms (just checking)


Former-commit-id: 07f963a98b8b8b8b7ed1048c02ba38b91a4739f9