Added more tests of mapped searches (all passed!)
Added a len to Cursor so that it is easier to loop over cursor sub-views

Re-ordered the enum for the different view sizes of searched objects, so that
the values were in view size order (smallest to largest)

Changed "all" and "*" to mean "the existing view", so that

>>> view["*"]

returns the current view (it is a null search) and

>>> view["all"]

does the same thing. This way, "bonds with *" and "bonds in *"
do what you expect. These are now distinct from "molecules",

>>> view["molecules"]

which is clearly an expansive search.

Added a description of what I mean by expansive and contractive
searches to the search cheat sheet.


Former-commit-id: 0ed186e3356b2c36ca8d9a623f9fe8bf775b0188