Cleaned up the grammar to remove the multi-line Expressions (it is now a single-line only, and kept in if I need to add some extra layers above).
Fixed various issues with the indexing grammar for complex indexing queries.

Played with using `( )` for the grouping syntax of subscript searches, but the `( )` got eaten by other grammar rules.
Ended us sticking with `{ }[index]` as this is easier.

Added tests for this, e.g.

`{element C in resname ALA}[1]`

also added tests for perturbable, e.g.

>>> assert mols["atoms with property atomtype == DU in molecules with property is_perturbable"] == mol[0]
>>> assert mols[("atoms with property atomtype == DU in molecules with property is_perturbable",
                {"atomtype": "final_atomtype"})] == mol[-1]


Former-commit-id: 1d91fa0327a19d684927506c5f8740a454746052