Fixed and added tests for property comparisons that have units - these now assume the default unit for the dimension being tested. This can be changed by the user.
This is tested and is also described in the search cheat sheet.

Also added the ability to set cursor properties to different values when
operating on a cursor collection. This is useful to set the coordinates
of multiple atoms at once.

This is described by an addition to the move tutorial, plus is
tested in `test_cursor_direct` in test_move.py. This works for
single-molecule and multiple-molecule cursor collections,
for any property.


Former-commit-id: 6fe95732e45823603a91a949ffaa422c4ec354f7