Made the error handling and warning reporting for Gromacs topology files better. Warnings are better captured (preventing thousands of repeat warnings), while also being better presented to the user as warnings. The warnings for a molecule are also now printed as part of any exception on file load, with the errors now captured in the right place so that we don't get confusing messages.
Once I did this, I then fixed the cause of the error - namely putting
the chain name into the residue name. I added support for this on
parsing, and then also added support to retain this information in
the parser and to write it out - meaning that the chain information
can now be read and written to a gromacs topology file.

Also fixed a precision error in the energy test - this was falsely
tripping a fail on Linux Python 3.10.


Former-commit-id: 1f858aea964897be1356961f827f7d5a4d126f07