Ok - got it. I use ExpressionPartRule for the left hand side (LHS) of any binary rule. ExpressionRule can be on the right hand side (RHS).
Then, I add an "|" to all "(" << ExpressionRule << ")" for the LHS,
meaning that now I don't get an infinite recursion as the parser
now won't be stuck infinitely parsing the LHS.

The brackets force the user to specify the precedence of the binary
searches (the expression must be in brackets to turn it from a
unary ExpressionPartRule to a complex (potentially binary
ExpressionRule). So I also need to add the options of brackets
around the RHS ExpressionRule, plus around the entire
binary expression. This worked :-)

Now I just need to re-work all of the other binary rules to
follow the same logic, and then I think they will all be
properly combinable without missing edge cases.

Still WIP, but now mostly working


Former-commit-id: a8f52895072a504fbc9d0624fbf17cd5cfd286f2