Ok - I've (finally) finished the first pass through the search cheat sheet.
This pretty much explains everything, except for the syntax around
searching with a specified property map. I will add that once
I work out how to explain property maps in the tutorial.

I am going to move on now to getting the pull request back to devel
ready, and then finishing the website for the 2023.1 release
next month.

Covid was annoying, but now I am recovering and have nothing to do
but isolate, it has at least given me some time to code.


Former-commit-id: 86fe305d4686cb94096b6ca5d4ec24a82228d52a