Finished the null bond/angle parameter fix. Null bonds and angles can now be written using their current lengths/sizes (default) or skipped entirely is `skip_null_bonds:False` or `skip_null_angles:False` is passed in the `map`.
Added some tests to check this. Found a bug in file saves that they weren't
being written to the correct directory if the format was specified.
Fixed this!

Also fixed the (very broken) `__copy__` and `__deepcopy__` functions in
`sire.system.System`.

Also documented how to use fixtures for reading files, plus how
to use `tmpdir` for writing files to a temporary directory
when writing tests in the developer guide.