Fixed the waterswap energy error and have added a test so it won't reappear.
The fix was very simple, but is masked by some autoformatting changes (black)

Error was that len(mol) was 1 in the old API. BUT the autoconversion from
ViewsOfMol to Python returned EITHER a Molecule or a list of views.
The original _molecule_join function assumed that there would only be a list
of views, as older versions of sire (pre 2023) would return a list of a
single Molecule in the single-Molecule case.

The join function ran len(view) on the Molecule, got `1` as the response,
and so returned only the first atom as the join. This first atom was
added to the forcefields, meaning that most of the protein intramolecular
energies were zero (or near zero, as they only contained a single atom).

The fix was to check if the object being joined is a list. If it is not,
then we assume it is a single view and return that (processed in the
same way as the list of views).

Writing the unit test to find this revealed how this got past QA.

This condition only appears using the old API, meaning all the
new API tests passed. I managed to create a test that faked
the old API, so that we will catch this if it ever
regresses in the future.

The other reason this got past QA was because adding one atom
did create some energies, and those energies were correct
(and were correctly calculated and book-keeped throughout
the simulation). They were just missing lots of
intramolecular terms! Short simulations that tested rough
free energies were thus sensible, and nothing got flagged up
because the impact of missing intramolecular protein terms
is only visible in a longer simulation.

I've checked and sire now gives the same energies for the
test waterswap system as 2022.1.0. I am running a longer
simulation now but think this should have fixed the bug.