Moved the calculate_energy functions to SireSystem, as they now need ForceFieldInfo to configure themselves.
Can now control if they are using shift electrostatics or reaction field.

Now integrating this with the sire_openmm converter, so that we can specify the cutoff type
and parameters during the conversion (or these can be sensibly guessed).

The last openmm unit test fails as I haven't yet finished the code to set the cutoff correctly.