Can now run minimisation on molecules. Have added code to extract coordinates and velocities from openmm and copy them back to sire. This looks to work well.
Also found that I had to make the `.update()` code more consistent. Now I can call
`.update()` on all containers / views and can pass in pretty much all containers
and views. This was a big update with little visible change, but it did hit
all of the containers and their wrappers (hence the large delta)

All the tests still pass :-)