Cleaned up the way the runner runs jobs - processing of blocks is also now in a background thread.
API now lets you specify the simulation in time to simulate, with save
frequency in time between saves. This is quite intuitive.

Added function-level arguments that can override / populate the
property map, so that the class is easier to use.

Made the progress bar transient, so it disappears after the run.

Added a str function for the Dynamics object that prints some
nice statistics after each run.