Re-activated the parallel code in AmberPrm - this looks like it works well and significantly speeds up loads and saves. Now, a system that took 20+ seconds to load or save completes in ~5 seconds.
I've also cleaned up the load and save functions so that the options
are easier to use (e.g. setting gromacs path, turning on and off
parallelisatin etc)

Also fixed a bug in AmberPrm where the number of dummies was not
passed correctly - instead the value of 'parallel' was passed
(0 or 1).

Now need to test this on Windows to see if I can find the crash
we had before.

I've tested this on Mac, and have also tested against the
files in https://github.com/michellab/sireparsertests/