Finished up the AmberTraj parser and have added a read/write unit test. It works well :-)
Also cleaned up how it handles infinite spaces, and added a check so that it only
supports Cartesian and PeriodicBox.

Fixed a small bug in AmberFormat when a PRM7 field is empty.

Added in the third party xdr support code from gromacs / mdanalysis (BSD-style license)
in preparation for writing the trr trajectory parser.