Lots of work to move everything across to a new ProgressBar/Spinner design.
This unifies the C++ and Python progress bars. I also added code to allow
long-running C++ functions to release the GIL. This is called from within
C++, and seems to be very stable. This is needed when using the progress
bar / spinner from within a C++ function when running in Jupyter,
as Jupyter needs time to process the C++ stdout to display it
in the cell output. Without this, Jupyter is blocked until after
C++ has finished doing whatever it was doing ;-)