Added a FileTrajectoryParser class that can be used to read in a directory of single-structure files that represent a complete trajectory. This complements the way we can now write out trajectories of single-frame files as a directory of files, e.g.
>>> f = sr.save(mols.trajectory(), "test", format=["PRMTOP", "RST7"])
>>> print(f)
["test.prmtop", "test.rst7"]

has saved the trajectory as a single PRMTOP file, plus a directory of
frames in RST7 format (the directory is called "test.rst7")

This can now be read back in, e.g.

>>> mols = sr.load("test.prmtop", "test.rst7")

The code recognises that "test.rst7" is a directory and scans this to find
all of the frame files. The files are then loaded only when they are needed
(thus saving a lot of time and memory).

In addition to writing the code to support this, I also ported
the trajectory energy code to use SireBase::ProgressBar entirely
within C++ - thus replacing the (messy) python code that dropped
in and out of C++ to update the Python bar, plus provide
Python with the opportunity to do things in parallel / send
interupts.

Now this code (like the trajectory writing code) releases the GIL
properly and is interuptable, so it should play well with Python
threads.