WIP - broken apart the old AmberRst parser and now reconstructing it from scratch based on the other trajectory parsers.
Have got the code working with extracting hyperslabs from netcdf files. This is a big improvment on the old
code that read the *whole* netcdf file into memory. This will let me stream the trajectory from the file.

Most of the foundations are in place now - just the work of actually loading and saving the file...