I have fixed the rounding / overflow errors caused by combining units by switching to a log10 algorithm when performing multiplication, division or raising units to a power.
The log10 helps cancel out large powers during the assembly of units, meaning that,
e.g. the very large values of ``gram``, ``mole`` and ``meter`` don't blow everything
up.

I have also added in code that creates nice strings from units, based on either
the user-set default unit (which is now much easier to set), or from the
defaults for the seven base physical units. I've added a unit test that
thoroughly tests this, and it works well.

I have added more documentation about this, and how it works, to the
units cheatsheet.