I've optimised the `makeWhole` function so that it now takes microseconds to ~4 ms to run on a system :-)
With this speed, it makes sense to add options that allow `makeWhole` to be called
automatically on system load, or when loading a trajectory frame.

I've added optimised functions that can make trajectory frames whole as they
are loaded into molecules (either at sr.load time or mols.load_frame time).

I've added a flag, "make_whole", which if True, will automatically
make molecules whole. This flag is default False (actually, default
False when it is not set). This way there is no change in
existing behaviour, and we can experiment a while with setting this flag
on lots of systems before deciding whether it should be automatically
switched on.

You use this flag via a `map={"make_whole": True}` argument to
`sr.load` or to `mols.load_frame`.

I've added some more tests that check that this works, and also
that check that this doesn't break or change existing systems
that don't need making whole.