Resolved hidden overloads so will hopefully remove many of the compiler warning messages that were flooding GH Actions.
Hopefully(?) fixed the issue with openmm failing thinking that it saw a
zero or negative periodicity. I can't reproduce this locally on any OS,
and have double-checked openmm versions. My only thought is that
I was using 'double' as the periodicity type when it should be int.

I've changed this to 'int', and have added checks that this is more
than 0. If it is 0, and k is 0, then I make sure to add the
dihedral with k 0 and periodicity 1.

Otherwise I print a warning message and drop the dihedral from the list.

I will check locally. I think this issue came up because conda now uses
gcc 13.1, but I have gcc 12 on my ARM64 Linux VM.