Added a "smarts X" search to the grammar. Can now capture the smarts string and pass it up and out to the SireRDKit code in wrapper/Convert. This required having the RDKit code register itself with the underlying SireSearch functionality.
Can now successfully return a molecule if any part matches the passed smarts string.

The next step is to only match the atoms in this molecule that match the smarts string.

Also, currently we have to do `mols["not water"]["smarts XXX"]` to remove the very
slow checking of all water molecules. It is really very slow to do substructure searching
on all waters...