Added `add_suffix` and `add_prefix` function to `PropertyMap` to make it easier to get properties from merged molecules, e.g.
```
>>> m = merged_molecule
>>> props = ["charge", "LJ", "bond", "angle" .... ]
>>> map = sr.base.create_map(map_passed_from_user)
>>> map0 = map.add_suffix("0", props)
>>> map1 = map.add_suffix("1", props)
```

will map `charge` to `charge0`, `LJ` to `LJ1` etc in `map0`, and
`charge` to `charge1` and `LJ` to `LJ1` etc in `map1`.

I've used this to easily extract the lambda=0 and lambda=1 parameters when
building the OpenMMMolecule (including making it easier to switch the
end states for testing).

Now looking at how these parameters can be mapped to forces.