Updated the `update_recipe.py` script to properly get the packages and versions, so that we only include each dependency once.
I've checked this now so that it does build a complete package :-)

Added a print out to `update_recipe.py` to give the `conda mambabuild`
command that includes any extra environments specified by
the `environment.yml`.

Also updated the docs to give more instructions, plus fixed the
`version.txt` for `main` so that it is the last build version,
not the `2023.4.0.dev` value.