Created a LambdaLever class (currently empty) that is created and attached to the OpenMM context, and accessible from Python via the `.get_lambda_lever()` function.
Also added on an index mapping from sire Atom to index in the OpenMM system, so that it is easier
to find and update atoms. This is also accessible from Python via a monkey-patched
function on the OpenMM context - `.get_atom_index()`.

Adding in the index made me get annoyed enough about the lack of + and += operators on
the moelcular container classes that I added them to all the T, Selector<T> and
SelectorM<T> types.