Lots of debugging and cleaning up of the API for AtomMapping. Needed to add a `.filter()` function to the `SelectorM<T>` container, as need to make sure I can find atoms that are in the container, returning the result in an order-preserving way.
This also added a `find_all` argument to the AtomMapping that tells it to either
match / find all atoms, or to mark missing atoms with `Atom()` (when mapped)
or don't return them when doing a `find()`.

This now works for mapped alignment of trajectory frames.

Now need to formalise all my test scripts into some proper unit tests.