Significantly sped up RMSDs and trajectory loading for large systems with lots of water molecules.
Did this by adding "updateProperty" and "updatePropertyFrom" functions that let molecule properties
be updated in place, thus avoiding copies and lots of memory reallocation.

Also added a "coords_only" flag that sets that only the coordinates of the molecules should
be loaded from the trajectory. This allows us to skip loading velocities, forces,
and space / time (the last two being really slow for some reason!) when we are only
interested in coordinates (e.g. when calculating RMSDs or showing movies)