I've fixed the positional restraint to be a proper flat-bottom well potential.
I've also added the restraints to the LambdaLever so that you can scale them
up and down with lambda. This is done in a type-agnostic way, meaning that
I can add lots of differernt types of restraints.

To fill these in, I have started work on two more restraint types;

1. BondRestraints (mostly done)
2. BoreschRestraints (todo)

Code compiles and runs