I've added the code to create and calculate the energies for the Boresch restraints.
You can create them using

```
>>> restraints = sr.mm.create_boresch_restraints(mols, receptor=receptor[0:3], ligand=ligand[0:3])
```

where you pass in three atoms of the receptor and three atoms of the ligand.

The code to automatically choose the best receptor and ligand atoms still needs
to be written.

The restraint has been added as a single CustomCompoundBondForce, using the
6 atoms that comprise the restraint.

This runs but needs tests, documentation and debugging