Call `takeProperty` on the trajectory, then save frame, then `setProperty`, rather than `property`, save frame, `setProperty` so that we don't thrash memory!
This way there is only a single copy of the trajectory at any one time,
so each append is not having to do a full copy. This was significantly
slowing down trajectories when large numbers of frames were saved!

Now the dynamics looks to have a consistent speed, at least on Mac/Metal