Made sure that we match dihedrals by atoms AND periodicity, so that we don't morph periodicity values
Added labels to the EnergyTrajectory so that we can store additional data, e.g.
the simulated lambda value. Also added the ability to save ensemble to the
System

This allows the `.energy_trajectory()` function to have a `to_alchemlyb`
option that outputs the energy trajectories as DataFrames in
an alchemlyb-compatible format.