Make sure that we don't record energies / frames when the update frequency is set to zero. This prevents energies / frames from being saved from minimisation if this has been disabled.
Caught the case when a system doesn't have a space, so automatically add a Cartesian space

Working on the free energy tutorial. Now run the simulation for a bit longer and
save to a s3 file instead of parquet. Added in the alchemlyb code to do the analysis,
although this gives the wrong result (Bennetts from sire gives the right result for this
data, so I think I have done something wrong with alchemlyb)