Added functions to let the user easily set the temperature and/or pressure, and also to trigger randomisation of the velocities.
Fixed a bug in OpenMMMolecule where it set the coordinates to 0,0,0 if the property
didn't exist in the molecule... Now is raises a missing_property exception as it should ;-)

Updated the tutorial to randomise velocities, and to use the correct free energies calculated
using the code as it is now. Randomising velocities has made the free energy slightly
different.

Also added a new section on how to make the simulation faster, where I show using the timestep
and constraints. Saw that the simulation is quite stable and that h-bond-h-atoms is the best
way to go, so have added this as an intermediate step.

Also shown how to use the hydrogen mass repartitioning function