Added the option to control whether or not the internal energy/forces of constrained degrees of freedom is included in the OpenMM context.
Exposed the "platform" option to the `.dynamics()` and `.minimisation()` functions.
Also removed their auto-parsing of kwargs, as this was a big source of typo
bugs when calling the function...

Fixed a bug in how the file formats were read in MoleculeParser where they came
only from the property map value, but should be checked for both the value
and the source. Updated the way create_map works so that it actually returns
when it has correctly added the value to the map, rather than keeping going
and always calling the expensive wrap function (hence why strings were
sometimes coming in as values rather than sources)

Exposed and tested all of this via a new test in test_openmm.py