I've changed the default hydrogen mass repartitioning (HMR) factor from 4 to 1.5, because 4 causes a LangevinMiddleIntegrator to become unstable with a 4fs timestep, but 1.5 seems fine. I've gone with the best default for the default integrator.
I've added a `vacuum` option to the dynamics and minimisation functions to make
it easier to specify that a simulation should be run in vacuum. This makes
the writing of vacuum legs easier, as now you don't need to manually set
the space.

I've finished off the free energy tutorial, adding an example script that
does the whole job of calculating a relative hydration free energy.