I think this fixes the compile error on Windows (missing definition of PhysUnit(GeneralUnit) constructor, because this is defined in generalunit.h, which is not included in dimensions.cpp)
I've also fixed a regression in MoleculeParser that was detected in sire_bigtests - the
code was not auto-setting the fileformat of a file that is written based on the fileformat
it had been loaded with.

Also removed the default `to_pandas=True` for `system.energy_trajectory()` as this was confusing,
and was from a time when I thought that the pandas dataframe could do everything. Now I know that
the missing functionality is needed, and provided by EnergyTrajectory. I've updated the tutorials
to remove this confusing argument.

This also blackened run_md.py

Let's see if this all compiles...