Explicitly instantiated AtomLJs so that I will be able to add functions to handle LJ_ij exceptions.
In doing this, I saw that I am not properly doing the magic and versioning for streaming
AtomProperty<T> objects...

Partly to fix this, I have added the check in SireStream::readHeader to put the stream back
to its last position if the magic error fails. This allows me to detect a wrong magic, and
to then choose a different class or try to stream differently. I need to check if this
causes any performance problems