I've implemented the full suite of `-not-perturbed` constraints, adding them to the documentation and options objects. Just running some simulations now to make sure they are working as expected. Debugging showed the right bonds and angles were not being constrained.
I've updated the defaults so that ``h-bonds-not-perturbed`` etc are the
default constraints if none are set. I've also added a discussion of PME
versus RF to the tutorial, and updated the example FEP script to
use ``h-bonds-not-perturbed`` and ``RF`` by default.

On my computer, the entire methane/ethanol FEP simulation now takes
about 3 minutes :-)