Done a lot of work on the new tutorial chapter that describes how to use the new functionality. Now working this up, updating the functionality as needed to make a nice tutorial.
There are some changes to the code which I've described in the changelog updates.

In particular, the behaviour of ``mol.perturbation().link_to_reference()``
has changed, meaning that you don't need to add a ``.commit()`` on the
end (indeed, it will now give an error).

The right fix for this is to use the new, higher level functions to
link to the reference or perturbed states, e.g.

```
mols = sr.morph.link_to_reference(mols)
```