I've added in a new constraint type - "X-not-heavy-perturbed". This constraints "X", but will not constrain perturbing bonds *unless* they involve a hydrogen in any end state.
I have switched this to be the default constraint type, as it is the closest match
to somd1. The real match would be
`constraint="h-bonds", perturbable_constraint="bonds-not-heavy-perturbed"`

I have also improved the code that detects "light" atoms (i.e. hydrogens).

This now uses the maximum mass in any end state being below 2.5, or if
an element in any end state is a H, or if the ambertype in any end state
matches "H*" (case-insensitive).

These three ways of matching are all switched on by default. They can
be disabled via options described in the OpenMM cheatsheet.