Added the ability in AtomMapping to store the atoms that weren't mapped at both end states. This means we now have the full information in this object to be able to properly construct merged and mutated molecules.
Added a unit test to verify this works for mapping an alanine to a lysine within a protein
(mapping works on subsets of molecules, meaning we can do a region of interest mapping
and merge just be passing in the residues that we want to map - see test_match.py)

Also, removed the C++ code for mutate as I realised that mutation is a merge
followed by extracting the perturbed state.

This is still a WIP - have got matching working, and have added the necessary info
to AtomMapping. The next step is to fill in the C++ code for merge based on the
information in AtomMapping.