WIP - VERY WIP
Gone a little down the rabbit hole in the merge code by adding in support for alternate
names for atoms and residues. This will enable the names of the perturbed end state
to be saved, so that we can get something sensible out when we extract the end states.

Will not compile