Changed `isDummy` to check epsilon values only.
Changed `fix_ghost_sigmas` into `fix_perturbable_zero_sigmas` and moved the test into
PerturbableOpenMMMolecule. Added ability to pass `map` to this constructor so that
the option can be passed through to the right place.

Most of the diff is because I had to regeneate the wrappers, and so the MM wrappers
all got the new copy<T>() functions for their copy and deepcopy functions.